Rename "ParsedClusterNetwork"/"networkInfo" to "SDNConfig"/"sdnConfig"
Originally common.ParsedClusterNetwork *just* contained the parsed
form of the ClusterNetworks/ServiceNetwork entries from
ClusterNetwork, and the naming around it reflected that. But
eventually it became a complete replacement for ClusterNetwork
internally, and as such basically just represents the SDN
configuration, so call the object "SDNConfig" and store it in
variables called "sdnConfig" as well.